third_party/WebKit/LayoutTests/platform/win/http/tests/push_messaging/subscribe-failure-mismatched-sender-id-expected.txt

This is a testharness.js-based test.
PASS subscribe() is rejected when a subscription with a different sender IDalready exists 
Harness: the test ran to completion.
